requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
requests==2.32.3
duckduckgo-search==6.2.11
beautifulsoup4==4.12.2